Using the AtomiClusterProvider instance in the tests
