+++ suuchi-cluster-atomix/src/main/resources/META-INF/services/in.ashwanthkumar.suuchi.cluster.ClusterProvider
@@ -1,0 +1,1 @@
+in.ashwanthkumar.suuchi.cluster.atomix.AtomixClusterProvider
